pyproject.toml
[project]
name = "fishometry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=6.0.0",
    "e3nn>=0.5.8",
    "einops>=0.8.1",
    "evo>=1.34.0",
    "fastapi>=0.122.0",
    "gradio>=6.0.2",
    "huggingface-hub>=1.1.5",
    "imageio>=2.37.2",
    "moviepy<2.0",
    "numpy<2",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pillow-heif>=1.1.1",
    "plyfile>=1.1.3",
    "polars>=1.35.2",
    "pyarrow>=22.0.0",
    "pycolmap>=3.13.0",
    "safetensors>=0.7.0",
    "segment-anything>=1.0",
    "streamlit>=1.52.1",
    "timm>=1.0.22",
    "tqdm>=4.67.1",
    "trimesh>=4.10.0",
    "typer>=0.20.0",
    "ultralytics>=8.3.232",
    "uvicorn>=0.38.0",
    "xformers>=0.0.33.post1",
]
